Massive overhaul!
Updated the generation and drawing code to pass around the current center x,y and the calculated tile size. This will allow live positioning and resizing.

Also made some minor tweaks to the way the board is drawn. It now has a transparent background and a thick black border (color can be easily adjusted).

Made the number of columns for the player bases easily changed in code.

When armies are submitted they are randomly placed around the ocean spaces now rather than lining up along the top. This may be final behavior for the game. Not sure yet.

Updated so that the princess die is added into each army for the capture the flag gamemode.

Updated some notes here and there, including adding/removing some TODO items.