Massive update. All basic functionality is in place.
Capable of detecting win conditions now.

Added the counter dice for the king of the hill gamemode.

Modified the cell effect evaluation method and added effects for the center square (but only for king of the hill gamemode).

Added board size dropdown and the ability to alter the number of base columns.

Pruned and modified some TODO notes.